+++ .github/workflows/perftest-pr-head.yml
@@ -13,65 +13,59 @@ jobs:
   build-and-upsert-comment:
     runs-on: ubuntu-latest
     steps:
-      - name: Checkout
-        uses: actions/checkout@v4
-
       - name: Checkout latest master
-        uses: actions/checkout@v4
+        uses: actions/checkout@v6
         with:
           ref: 'master'
           path: master
 
       - name: Checkout pr head
-        uses: actions/checkout@v4
+        uses: actions/checkout@v6
         with:
           ref: ${{ github.event.pull_request.head.sha }}
           path: pr-head
-
-      - name: Configure Git
-        run: |
-          git config user.name "GitHub Actions Bot"
-          git config user.email "actions@github.com"
 
       - name: Setup Cpp
         uses: aminya/setup-cpp@v1
         with:
           cmake: true
+          ninja: true
 
       - name: Build latest master and PR head
         id: build
         run: |
-          mkdir -p ../builds
+          mkdir -p builds
 
-          cmake -B ../builds/master -S master -DFINUFFT_BUILD_TESTS=ON &&
-          cmake --build ../builds/master --target perftest -j
+          cmake -B builds/master -S master -DFINUFFT_BUILD_TESTS=ON -DCMAKE_BUILD_TYPE=Release &&
+          cmake --build builds/master --target perftest -j
 
           rm -fr pr-head/perftest
           cp -r master/perftest pr-head/perftest
-          cmake -B ../builds/pr-head -S pr-head -DFINUFFT_BUILD_TESTS=ON &&
-          cmake --build ../builds/pr-head --target perftest -j
+          cmake -B builds/pr-head -S pr-head -DFINUFFT_BUILD_TESTS=ON -DCMAKE_BUILD_TYPE=Release &&
+          cmake --build builds/pr-head --target perftest -j
 
           pip install pandas matplotlib
           python3 master/perftest/perftest_pr_head.py \
-            --master-perftest ../builds/master/perftest/perftest \
-            --pr-perftest ../builds/pr-head/perftest/perftest \
-            --plot-output perftest_comment_figure.png
+            --master-perftest builds/master/perftest/perftest \
+            --pr-perftest builds/pr-head/perftest/perftest \
+            --plot-output master/perftest_comment_figure.svg
 
       - name: Commit benchmark results
         id: push-image
         env:
           GITHUB_TOKEN: ${{ github.token }}
         run: |
+          cd master
+          git config user.name "GitHub Actions Bot"
+          git config user.email "actions@github.com"
           git branch -D perftest-comment-figure || true && git switch --orphan perftest-comment-figure
-          git add perftest_comment_figure.png
+          git add perftest_comment_figure.svg
           PRE_COMMIT_ALLOW_NO_CONFIG=1 git commit -m "Performance change [no ci]"
           IMAGE_SHA=$(git rev-parse HEAD)
-          echo $IMAGE_SHA
           git push -f https://${GITHUB_TOKEN}@github.com/${{ github.repository }} perftest-comment-figure
           echo "image_sha=$IMAGE_SHA" >> $GITHUB_OUTPUT
 
       - name: Find existing managed comment
-        if: always()
         id: find-comment
         uses: peter-evans/find-comment@v3
         with:
@@ -80,20 +74,20 @@ jobs:
           body-includes: "<!-- pr-managed-comment -->"
 
       - name: Create managed PR comment
-        if: always() && steps.find-comment.outputs.comment-id == ''
+        if: steps.find-comment.outputs.comment-id == ''
         uses: peter-evans/create-or-update-comment@v4
         with:
           issue-number: ${{ github.event.pull_request.number }}
           body: |
             <!-- pr-managed-comment -->
-            ![Perftest plot](https://raw.githubusercontent.com/${{ github.repository }}/${{ steps.push-image.outputs.image_sha }}/perftest_comment_figure.png)
+            ![Perftest plot](https://raw.githubusercontent.com/${{ github.repository }}/${{ steps.push-image.outputs.image_sha }}/perftest_comment_figure.svg)
 
       - name: Update managed PR comment
-        if: always() && steps.find-comment.outputs.comment-id != ''
+        if: steps.find-comment.outputs.comment-id != ''
         uses: peter-evans/create-or-update-comment@v4
         with:
           comment-id: ${{ steps.find-comment.outputs.comment-id }}
           edit-mode: replace
           body: |
             <!-- pr-managed-comment -->
-            ![Perftest plot](https://raw.githubusercontent.com/${{ github.repository }}/${{ steps.push-image.outputs.image_sha }}/perftest_comment_figure.png)
+            ![Perftest plot](https://raw.githubusercontent.com/${{ github.repository }}/${{ steps.push-image.outputs.image_sha }}/perftest_comment_figure.svg)
