+++ perftest/perftest_pr_head.py
@@ -1,67 +1,26 @@
 import argparse
 import subprocess
 from datetime import datetime
-from dataclasses import dataclass, fields
-from numbers import Number
 import io
+
 import pandas as pd
 import matplotlib.pyplot as plt
 import numpy as np
 
-
-@dataclass(frozen=True)
-class Params:
-    prec: str = "f"
-    N1: Number = 320
-    N2: Number = 320
-    N3: Number = 1
-    ntransf: int = 1
-    threads: int = 1
-    M: Number = 1e6
-    tol: float = 1e-5
-
-    def ndim(self) -> int:
-        if self.N3 > 1:
-            return 3
-        if self.N2 > 1:
-            return 2
-        return 1
-
-    def args(self) -> list[str]:
-        return [f"{f.name}={getattr(self, f.name)}" for f in fields(self)]
-
-    def pretty_string(self) -> str:
-        n = 4
-        fvalues = [f"{f.name}:{getattr(self, f.name)}" for f in fields(self)]
-        chunks = [" ".join(fvalues[i : i + n]) for i in range(0, len(fvalues), n)]
-        return "\n".join(chunks)
-
-
-NRUNS = 10
-
-PARAM_LIST = [
-    Params("f", 1e2, 1, 1, 1, 1, 1e7, 1e-4),
-    Params("f", 320, 320, 1, 1, 1, 1e7, 1e-5),
-    Params("d", 320, 320, 1, 1, 1, 1e7, 1e-9),
-    Params("f", 320, 320, 1, 1, 0, 1e7, 1e-5),
-    Params("d", 192, 192, 128, 1, 0, 1e7, 1e-7),
-]
-
-TRANSFORMS = [1, 2, 3]
-
-
-DEFAULT_EXTRA_ARGS = [
-    f"n_runs={NRUNS}",
-    "sort=1",
-    "upsampfact=0",
-    "kerevalmethod=1",
-    "debug=0",
-    "bandwidth=1.0",
-]
+from perftest_config import PARAM_LIST, NRUNS, TRANSFORMS
 
 
 def run_one_perftest(param, transform, binary_path: str) -> pd.DataFrame:
-    cmd = [binary_path] + param.args() + DEFAULT_EXTRA_ARGS + [f"type={transform}"]
+
+    extra_args = [
+        f"n_runs={NRUNS}",
+        "sort=1",
+        "upsampfact=0",
+        "kerevalmethod=1",
+        "debug=0",
+        "bandwidth=1.0",
+    ]
+    cmd = [binary_path] + param.args() + extra_args + [f"type={transform}"]
     result = subprocess.run(
         cmd,
         stdout=subprocess.PIPE,
